Generalize JsonLibrary interface for collection types
The previous iteration had a specialized deserializeListAt method to deserialize a list of the given class. This changes deserializeAt to take java.lang.reflect.Type instead of java.lang.Class to specify the type the caller wants to deserialize to. This allows the caller to specify parameterized types such as List as well. This enables support for many other types such as Collection without further code changes. Overall, the amount of implementation code required decreased at the same time.
Tests and docs have been updated and added where appropiate.