Fix error handling for List payload/return values
Both cases used crude placeholder handling for errors. Since the handling for `List`s is now very similar to POJOs, the code moved into the same block as POJOs and shares the errorhandling.

Also fixes a bug there the payload type was used for the return value when a `List` was used.